slack_eventsディレクトリ、parrot_bot.py作成し、ファイルを分割 #4
