+++ apis/routers/develop.py
@@ -1,3 +1,5 @@
+# developは、DB接続等の動作確認のためのエンドポイントをまとめています（本番環境では不要です）
+
 from typing import Annotated
 from fastapi import APIRouter, status, Depends
 from sqlalchemy import select
@@ -21,7 +23,7 @@ def read_root():
     "/develop/contents", response_model=ContentResponse, status_code=status.HTTP_201_CREATED
     )
 async def add_content(db: DbDependency, request: AddContentRequest):
-    # データ作成（id, content_date, is_deliveredはDB側でデフォルト値が入る）
+    # データ作成（id, content_date, is_deliveredは登録時にDB側でデフォルト値が入る）
     new_content = Content(
         team_id = request.team_id,
         user_id = request.user_id,
@@ -41,6 +43,8 @@ async def add_content(db: DbDependency, request: AddContentRequest):
     "/develop/contents/team/{team_id}", response_model=list[ContentResponse], status_code=status.HTTP_200_OK
     )
 async def get_contents(db: DbDependency, team_id):
+    # select()内の項目が複数の場合はdb.execute…を使う。
+    # （scalarsは最初の項目しか取得できないため）
     result = await db.scalars(select(Content).filter(Content.team_id == team_id))
     contents = result.all()
